fix: separate UI and main to different files
