+++ packages/main/src/plugin/authentication.spec.ts
@@ -23,37 +23,17 @@ import type {
   AuthenticationSessionAccountInformation,
   Event,
 } from '@podman-desktop/api';
-import type { BrowserWindow } from 'electron';
-import { Menu } from 'electron';
 import { afterEach, beforeEach, expect, test, vi } from 'vitest';
 
 import type { ApiSenderType } from './api.js';
 import { AuthenticationImpl } from './authentication.js';
 import { Emitter as EventEmitter } from './events/emitter.js';
 
-const menu = {
-  popup: vi.fn(),
-} as unknown as Menu;
-
-vi.mock('electron', () => ({
-  Menu: {
-    buildFromTemplate: (): Menu => menu,
-  },
-  BrowserWindow: {
-    getFocusedWindow: (): BrowserWindow =>
-      ({
-        webContents: {
-          getZoomFactor: (): number => 1.3,
-        },
-      }) as unknown as BrowserWindow,
-  },
-}));
-
 function randomNumber(n = 5): number {
   return Math.round(Math.random() * 10 * n);
 }
 
-class RandomAuthenticationSession implements AuthenticationSession {
+export class RandomAuthenticationSession implements AuthenticationSession {
   id: string;
   accessToken: string;
   account: AuthenticationSessionAccountInformation;
@@ -67,10 +47,11 @@ class RandomAuthenticationSession implements AuthenticationSession {
   }
 }
 
-class AuthenticationProviderSingleAccount implements AuthenticationProvider {
+export class AuthenticationProviderSingleAccount implements AuthenticationProvider {
   private _onDidChangeSession = new EventEmitter<AuthenticationProviderAuthenticationSessionsChangeEvent>();
   private session: AuthenticationSession | undefined;
   onDidChangeSessions: Event<AuthenticationProviderAuthenticationSessionsChangeEvent> = this._onDidChangeSession.event;
+  constructor() {}
   async getSessions(scopes?: string[]): Promise<readonly AuthenticationSession[]> {
     if (scopes) {
       return this.session ? [this.session] : [];
@@ -274,67 +255,4 @@ test('authentication provider send event to update settings page', async () => {
   disposable.dispose();
 
   expect(apiSender.send).lastCalledWith('authentication-provider-update', { id: 'provider1.id' });
-});
-
-test('showAccountsMenu creates menu with authentication requests and current sessions', async () => {
-  const authProvidrer1 = new AuthenticationProviderSingleAccount();
-  const authProvidrer2 = new AuthenticationProviderSingleAccount();
-  authModule.registerAuthenticationProvider('company.auth-provider1', 'Provider 1', authProvidrer1);
-  authModule.registerAuthenticationProvider('company.auth-provider2', 'Provider 2', authProvidrer2);
-
-  const session1 = await authModule.getSession(
-    { id: 'ext1', label: 'Ext 1' },
-    'company.auth-provider2',
-    ['scope1', 'scope2'],
-    { createIfNone: true },
-  );
-  expect(session1).toBeDefined();
-
-  const session2 = await authModule.getSession(
-    { id: 'ext2', label: 'Ext 2' },
-    'company.auth-provider1',
-    ['scope1', 'scope2'],
-    { silent: false },
-  );
-  expect(session2).toBeUndefined();
-
-  const session3 = await authModule.getSession(
-    { id: 'ext3', label: 'Ext 3' },
-    'company.auth-provider1',
-    ['scope1', 'scope2'],
-    { silent: false },
-  );
-  expect(session3).toBeUndefined();
-
-  expect(authModule.getSessionRequests()).length(2);
-
-  const buildFromTemplateSpy = vi.spyOn(Menu, 'buildFromTemplate');
-
-  await authModule.showAccountsMenu(10, 10);
-
-  expect(menu.popup).toBeCalledWith({ x: 13, y: 13 });
-  expect(buildFromTemplateSpy).toBeCalledWith(
-    expect.arrayContaining([
-      expect.objectContaining({
-        label: 'Manage authentication',
-      }),
-      expect.objectContaining({
-        type: 'separator',
-      }),
-      expect.objectContaining({
-        label: 'Sign in with Provider 1 to use Ext 2',
-      }),
-      expect.objectContaining({
-        label: 'Sign in with Provider 1 to use Ext 3',
-      }),
-      expect.objectContaining({
-        label: `${session1?.account.label} (Provider 2)`,
-        submenu: expect.arrayContaining([
-          expect.objectContaining({
-            label: 'Sign out',
-          }),
-        ]),
-      }),
-    ]),
-  );
 });
