+++ packages/main/src/plugin/index.ts
@@ -101,6 +101,7 @@ import type { WebviewInfo } from './api/webview-info.js';
 import { AppearanceInit } from './appearance-init.js';
 import type { AuthenticationProviderInfo } from './authentication.js';
 import { AuthenticationImpl } from './authentication.js';
+import { showAccountsMenu } from './authentication-menu.js';
 import { AutostartEngine } from './autostart-engine.js';
 import { CancellationTokenRegistry } from './cancellation-token-registry.js';
 import { Certificates } from './certificates.js';
@@ -1469,7 +1470,7 @@ export class PluginSystem {
     );
 
     this.ipcHandle('authentication:showAccountsMenu', async (_listener, x: number, y: number): Promise<void> => {
-      return authentication.showAccountsMenu(x, y);
+      return showAccountsMenu(x, y, authentication, apiSender);
     });
 
     this.ipcHandle(
